alert build process
