+++ src/main.js
@@ -1,0 +1,9 @@
+import alertFn from './alert';
+import confirmFn from './confirm';
+
+const alertjs = {
+	alert: alertFn,
+	confirm: confirmFn,
+};
+
+export default alertjs;
